game-flow: fix saving with disable-extra-guns
Resolves #4178.
