+++ src/libtrx/game/game_flow/util.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void GF_DisableObjectsIfNeeded(void);
