+++ src/libtrx/game/level/common.c
@@ -1,8 +1,6 @@
 #include "benchmark.h"
 #include "config.h"
-#include "debug.h"
 #include "game/camera.h"
-#include "game/creature.h"
 #include "game/effects.h"
 #include "game/game.h"
 #include "game/game_string_table.h"
@@ -17,58 +15,7 @@
 #include "game/random.h"
 #include "game/savegame.h"
 #include "game/sound.h"
-#include "game/viewport.h"
 #include "log.h"
-
-static void M_DisableObject(const OBJECT_ID object_id)
-{
-    OBJECT *const obj = Object_Get(object_id);
-    obj->loaded = false;
-    obj->collision_func = nullptr;
-    obj->control_func = nullptr;
-    obj->draw_func = nullptr;
-    obj->floor_height_func = nullptr;
-    obj->ceiling_height_func = nullptr;
-}
-
-static void M_ReplaceObject(
-    const OBJECT_ID src_object_id, const OBJECT_ID dst_object_id)
-{
-    for (int32_t i = 0; i < Item_GetTotalCount(); i++) {
-        ITEM *const item = Item_Get(i);
-        if (item->object_id == src_object_id) {
-            item->object_id = dst_object_id;
-        }
-    }
-}
-
-static void M_DisableObjectsIfNeeded(void)
-{
-    const GF_LEVEL *const level = Game_GetCurrentLevel();
-    if (level == nullptr
-        || (level->type != GFL_NORMAL && level->type != GFL_BONUS
-            && level->type != GFL_GYM)) {
-        return;
-    }
-    if (g_Config.gameplay.disable_medpacks) {
-        M_DisableObject(O_SMALL_MEDIPACK_ITEM);
-        M_DisableObject(O_LARGE_MEDIPACK_ITEM);
-    }
-
-    if (g_Config.gameplay.disable_extra_guns) {
-        const RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
-        ASSERT(resume != nullptr);
-        for (int32_t i = 0; g_GunObjects[i] != NO_OBJECT; i++) {
-            if (resume->flags.has_pistols) {
-                M_DisableObject(g_GunObjects[i]);
-            } else {
-                M_ReplaceObject(g_GunObjects[i], O_PISTOL_ITEM);
-            }
-            M_DisableObject(
-                Object_GetCognate(g_GunObjects[i], g_GunAmmoObjectMap));
-        }
-    }
-}
 
 void Level_Unload(void)
 {
@@ -139,7 +86,6 @@ bool Level_Initialise(
 
     Level_Unload();
     Level_Load(level);
-    M_DisableObjectsIfNeeded();
 
     GameStringTable_Apply(level);
